@@ -1,18 +1,19 @@
 {
   "scripts": {
-    "start": "vite",
+    "start": "vite --open",
     "build": "tsc && vite build",
     "preview": "vite preview"
   },
   "dependencies": {
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "@xyflow/react": "12.4.4",
     "ts-node": "10.3.0"
   },
   "devDependencies": {
     "@types/node": "16.11.0",
-    "@types/react": "17.0.30",
-    "@types/react-dom": "17.0.9",
+    "@types/react": "^18.2.0",
+    "@types/react-dom": "^18.2.0",
     "@vitejs/plugin-react": "^4.2.1",
     "typescript": "5.4.3",
     "vite": "^5.1.6"
